+++ Source/Canvas.h
@@ -112,7 +112,7 @@ public:
    
 private:
    void OnClicked(int x, int y, bool right) override;
-   void GetDimensions(int& width, int& height) override { width = mWidth; height = mHeight; }
+   void GetDimensions(float& width, float& height) override { width = mWidth; height = mHeight; }
    ofVec2f RescaleForZoom(float x, float y) const;
    
    bool ShowVerticalScrollBar() const;
