+++ Source/ChordDisplayer.h
@@ -30,6 +30,6 @@ public:
 private:
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(int& width, int& height) override { width = 200; height = 20; }
+   void GetModuleDimensions(float& width, float& height) override { width = 200; height = 20; }
    bool Enabled() const override { return true; }
 };
