+++ Source/CircleSequencer.h
@@ -76,7 +76,7 @@ public:
 private:
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(int& width, int& height) override { width=400; height=200; }
+   void GetModuleDimensions(float& width, float& height) override { width=400; height=200; }
    void OnClicked(int x, int y, bool right) override;
    
    std::vector<CircleSequencerRing*> mCircleSequencerRings;
