+++ Source/DelayEffect.h
@@ -53,7 +53,7 @@ public:
    
 private:
    //IDrawableModule
-   void GetModuleDimensions(int& width, int& height) override { width = mWidth; height = mHeight; }
+   void GetModuleDimensions(float& width, float& height) override { width = mWidth; height = mHeight; }
    void DrawModule() override;
    
    float GetMinDelayMs() const;
