+++ Source/FeedbackModule.h
@@ -42,7 +42,7 @@ public:
 private:
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(int& w, int&h) override { w=120; h=80; }
+   void GetModuleDimensions(float& w, float& h) override { w=120; h=80; }
    bool Enabled() const override { return mEnabled; }
    
    DelayEffect mDelay;
