+++ Source/HelpDisplay.h
@@ -30,7 +30,7 @@ private:
    //IDrawableModule
    void DrawModule() override;
    bool Enabled() const override { return true; }
-   void GetModuleDimensions(int& w, int& h) override;
+   void GetModuleDimensions(float& w, float& h) override;
    
    void LoadHelp();
