+++ Source/Lissajous.h
@@ -42,7 +42,7 @@ public:
 private:
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(int& w, int&h) override { w=mWidth; h=mHeight; }
+   void GetModuleDimensions(float& w, float& h) override { w=mWidth; h=mHeight; }
    bool Enabled() const override { return mEnabled; }
    
    float mWidth;
