+++ Source/LiveGranulator.h
@@ -49,7 +49,7 @@ private:
    
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(int& w, int& h) override { w = mWidth; h = mHeight; }
+   void GetModuleDimensions(float& w, float& h) override { w = mWidth; h = mHeight; }
    bool Enabled() const override { return mEnabled; }   
    
    float mBufferLength;
