+++ Source/Metronome.h
@@ -47,7 +47,7 @@ private:
    //IDrawableModule
    void DrawModule() override;
    bool Enabled() const override { return mEnabled; }
-   void GetModuleDimensions(int& width, int& height) override { width=80; height=35; }
+   void GetModuleDimensions(float& width, float& height) override { width=80; height=35; }
 
    float mPhase;
    float mPhaseInc;
