+++ Source/ModulatorAddCentered.h
@@ -46,7 +46,7 @@ public:
 private:
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(int& w, int&h) override { w=106; h=17*3+4; }
+   void GetModuleDimensions(float& w, float& h) override { w=106; h=17*3+4; }
    bool Enabled() const override { return mEnabled; }
    
    float mValue1;
