+++ Source/ModwheelToVibrato.h
@@ -39,7 +39,7 @@ public:
 private:
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(int& width, int& height) override { width = 138; height = 22; }
+   void GetModuleDimensions(float& width, float& height) override { width = 138; height = 22; }
    bool Enabled() const override { return mEnabled; }
    
    NoteInterval mVibratoInterval;
